requirements.txt
# ==========================================
# 🧠 CORE AI & ORCHESTRATION
# ==========================================
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-sdk==0.3.12
langchain==1.2.15
langchain-core==1.2.27
langchain-community==0.4.1
langchain-google-genai==4.2.1
google-generativeai==0.8.6
langsmith==0.7.26

# ==========================================
# 🏗️ DATA VALIDATION & SETTINGS
# ==========================================
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
annotated-types==0.7.0

# ==========================================
# 🌐 DATA ACQUISITION (Scraping & API)
# ==========================================
firecrawl-py==4.22.1
beautifulsoup4==4.14.3
httpx==0.28.1
requests==2.33.1
aiohttp==3.13.5

# ==========================================
# 🎨 UI & EXPORT
# ==========================================
streamlit==1.56.0
fpdf2==2.7.8
pandas==3.0.2
numpy==2.4.4
python-dotenv==1.2.2

# ==========================================
# 🛠️ UTILITIES & BACKGROUND (Auto-Generated)
# ==========================================
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
altair==6.0.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.6
dataclasses-json==0.6.7
distro==1.9.0
filetype==1.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonschema==4.26.0
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
nest-asyncio==1.6.0
orjson==3.11.8
packaging==26.0
pillow==12.2.0
protobuf==5.29.6
pyarrow==23.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.49
tenacity==9.1.4
tornado==6.5.5
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0